mesh: consolidate shared utilities into routers/comm/
Move error, header_utils, persistence_utils, conversations, parse,
and tokenize into comm/ (formerly shared/) — groups all cross-cutting
utilities and simple handlers under one module.
