+++ atom/mesh/src/routers/http_router.rs
@@ -36,9 +36,9 @@ use crate::{
         responses::{ResponsesGetParams, ResponsesRequest},
     },
     routers::{
-        error::{self, extract_error_code_from_response},
-        header_utils,
-        shared::{
+        comm::error::{self, extract_error_code_from_response},
+        comm::header_utils,
+        comm::{
             metrics_utils::{error_type_from_status, route_to_endpoint},
             placement_response::placement_err_to_response,
         },
