+++ atom/mesh/src/routers/openai/responses/persistence.rs
@@ -11,7 +11,7 @@ use crate::{
         common::Tool,
         responses::{ResponseTool, ResponseToolType, ResponsesRequest, ResponsesResponse},
     },
-    routers::persistence_utils::persist_conversation_items,
+    routers::comm::persistence_utils::persist_conversation_items,
 };
 
 /// Extract function tools from ResponseTools
