--- atom/mesh/src/routers/shared/mod.rs
@@ -1,2 +1,0 @@
-pub mod metrics_utils;
-pub mod placement_response;
